Fix directory permission issue in the MB pack
Before this fix some directories including the parent directory was given read
write permission to all the users (i.e. 777). Adding the version to
maven-assembly-plugin fixed the issue. But after the fix axis2.xml
was added to the MB zip twice building rpoduct zip. The extra copying was
excluded in bin.xml to avoid this. Also the version number of log4j was removed
from the distribution/pom.xml. Now the version is inherited from parent pom and
therefore log4j version conflicts are avoided.